Fix compile of Power.TonOfRefrigeration, use double not decimal
fixup! 4d89626 - Merge remote-tracking branch 'origin/master' into release/v6

+++ UnitsNet.Tests/CustomCode/PowerTests.cs
@@ -26,7 +26,7 @@ namespace UnitsNet.Tests
 
         protected override double DeciwattsInOneWatt => 1e1;
 
-        protected override decimal TonsOfRefrigerationInOneWatt => 2.8434512332474516279184828026648e-4m;
+        protected override double TonsOfRefrigerationInOneWatt => 2.8434512332474516279184828026648e-4;
 
         protected override double WattsInOneWatt => 1;
